Cleans up redundant code in proxy/internal RAPIDS Shuffle Managers (#6030)
